+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Add `ise_downloadable_acl` resource and data source
 - Add `ise_tacacs_command_set` resource and data source
+- Add `ise_tacacs_profile` resource and data source
 
 ## 0.1.5